# Add performance tests with its make target