--- a/Makefile
+++ b/Makefile
@@ -58,7 +58,7 @@ build-client:
 release:
 	./hack/release.sh
 
-.PHONY: image 
+.PHONY: image
 image: GOOS := linux
 image: .hack-apiserver-image .hack-client-image
 
@@ -90,7 +90,7 @@ go-fmt:
 	@echo 'Formatting go code...'
 	@gofmt -s -w .
 	@echo 'Not formatiing issues found in go codebase!'
-	
+
 
 .PHONY: check-license
 check-license:
@@ -145,7 +145,19 @@ test-e2e:
 	@. local/.env.local && go test -race github.com/adobe/cluster-registry/test/e2e -count=1 -v
 	$(shell pwd)/local/cleanup.sh
 
+IMAGE_TAG_PERFORMENCE=latest
+IMAGE_PERFORMENCE_TESTS=performance:${IMAGE_TAG_PERFORMENCE}
+API_URL=http://cluster-registry-api:8080
+NETWORK=cluster-registry-net
+NR_OF_SECS=10
 
+build-performance:
+	@echo "Building docker image '${IMAGE_PERFORMENCE_TESTS}'..."
+	@docker build -q -t ${IMAGE_PERFORMENCE_TESTS} -f test/performance/Dockerfile .
+	@printf "Done!\n\n"
+
+test-performance: build-performance
+	@docker run --rm --network=$(NETWORK) performance ${API_URL} ${NR_OF_SECS} ${API_TOKEN}
 
 ###############
 # Development #