--- a/pkg/apiserver/web/version.go
+++ b/pkg/apiserver/web/version.go
@@ -1,0 +1,25 @@
+/*
+Copyright 2021 Adobe. All rights reserved.
+This file is licensed to you under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License. You may obtain a copy
+of the License at http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software distributed under
+the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+OF ANY KIND, either express or implied. See the License for the specific language
+governing permissions and limitations under the License.
+*/
+
+package web
+
+import (
+	"net/http"
+
+	"github.com/labstack/echo/v4"
+)
+
+// Version return an handler that returns the version of the app
+func Version(version string) echo.HandlerFunc {
+	return func(c echo.Context) error {
+		return c.String(http.StatusOK, version)
+	}
+}